definition change for strhsh

+++ inc/strhsh.h
@@ -8,6 +8,6 @@ unsigned long long int getHashValue(const char* s);
 
 unsigned long long int getHashValueDstring(const dstring* dstr);
 
-unsigned long long int getHashValueByLength(const char* s,int length);
+unsigned long long int getHashValueByLength(const char* s, unsigned long long int length);
 
 #endif
